+++ src/angular-maps/services/geo.loader.ts
@@ -1,26 +1,9 @@
-import {GeoElement, GeoParser} from './geo.parser';
+import {GeoParser} from './geo.parser';
 import {MinimalHttpBackendClient} from '@dps/mycms-commons/dist/commons/services/minimal-http-backend-client';
-import {LogUtils} from '@dps/mycms-commons/dist/commons/utils/log.utils';
+import {CommonGeoLoader} from '@dps/mycms-commons/dist/geo-commons/services/geo.loader';
 
-export class GeoLoader  {
-    constructor(private http: MinimalHttpBackendClient, private parser: GeoParser) {}
-
-    loadDataFromUrl(url: string, options): Promise<GeoElement[]> {
-        const me = this;
-        return new Promise<GeoElement[]>((resolve, reject) => {
-            me.http.makeHttpRequest({ method: 'get', url: url, withCredentials: true })
-                .then(function onLoaded(res: any) {
-                    return resolve(me.parser.parse(res.text(), options));
-            }).catch(function onError(error: any) {
-                    console.error('loading geofeature failed:' + LogUtils.sanitizeLogMsg(url), error);
-                    return reject(error);
-            });
-        });
-    }
-
-    loadData(src: string, options): Promise<GeoElement[]> {
-        return new Promise<GeoElement[]>((resolve) => {
-            return resolve(this.parser.parse(src, options));
-        });
+export class GeoLoader extends CommonGeoLoader  {
+    constructor(http: MinimalHttpBackendClient, parser: GeoParser) {
+        super(http, parser);
     }
 }