+++ src/angular-maps/services/geo.parser.ts
@@ -1,14 +1,14 @@
-import * as L from 'leaflet';
-import LatLng = L.LatLng;
+import {LatLng} from 'leaflet';
+import {
+    GeoElementBase,
+    GeoElementType,
+    LatLngTimeBase
+} from '@dps/mycms-commons/dist/geo-commons/model/geoElementTypes';
+import {AbstractGeoParser} from '@dps/mycms-commons/dist/geo-commons/services/geo.parser';
 
-export enum GeoElementType {
-    TRACK,
-    ROUTE,
-    WAYPOINT,
-    AREA
-}
+export { GeoElementType as GeoElementType };
 
-export class LatLngTime extends LatLng {
+export class LatLngTime extends LatLng implements LatLngTimeBase {
     time: Date;
     constructor(latitude: number, longitude: number, altitude: number, time: Date) {
         super(latitude, longitude, altitude);
@@ -16,37 +16,17 @@ export class LatLngTime extends LatLng {
     }
 }
 
-export class GeoElement {
+export class GeoElement implements GeoElementBase<LatLng> {
     type: GeoElementType;
     points: LatLng[] = [];
     name: string;
 
-    constructor(type: GeoElementType, points: L.LatLng[], name: string) {
+    constructor(type: GeoElementType, points: LatLng[], name: string) {
         this.type = type;
         this.points = points;
         this.name = name;
     }
 }
 
-export abstract class GeoParser  {
-    abstract parse(src: string, options): GeoElement[];
-
-    _humanLen(l) {
-        if (l < 2000) {
-            return l.toFixed(0) + ' m';
-        } else {
-            return (l / 1000).toFixed(1) + ' km';
-        }
-    }
-
-    _polylineLen(ll: LatLng[]) {
-        let d = 0, p = null;
-        for (let i = 0; i < ll.length; i++) {
-            if (i && p) {
-                d += p.distanceTo(ll[i]);
-            }
-            p = ll[i];
-        }
-        return d;
-    }
+export interface GeoParser extends AbstractGeoParser<LatLng> {
 }