+++ src/angular-maps/services/geo.utils.ts
@@ -1,6 +1,6 @@
-import * as L from 'leaflet';
+import {LatLng} from 'leaflet';
 import * as tzlookup from 'tz-lookup/tz';
-import LatLng = L.LatLng;
+import {GeoElement, GeoElementType, LatLngTime} from './geo.parser';
 
 export class GeoUtils  {
     public static getLocalDateTimeForLatLng(pos: LatLng): Date {
@@ -17,4 +17,15 @@ export class GeoUtils  {
         return new Date(dateString);
     }
 
+    public static createLatLng(lat: string | number, lng: string | number, alt?: number, time?: Date): LatLng {
+        return time !== undefined
+            ? new LatLngTime(Number(lat), Number(lng), Number(alt), time)
+            : alt !== undefined
+                ? new LatLng(Number(lat), Number(lng), Number(alt))
+                : new LatLng(Number(lat), Number(lng));
+    }
+
+    public static createGeoElement(type: GeoElementType, points: LatLng[], name: string): GeoElement {
+        return new GeoElement(type, points, name)
+    }
 }